Concat 1st and 3rd party base js
