+++ package.json
@@ -16,7 +16,10 @@
     "flatpickr": "^4.6.6",
     "fullcalendar": "^5.3.2",
     "gulp": "^4.0.2",
+    "gulp-concat": "^2.6.1",
     "gulp-flatten": "^0.4.0",
+    "gulp-if": "^3.0.0",
+    "gulp-ignore": "^3.0.0",
     "gulp-postcss": "^8.0.0",
     "gulp-sass": "^4.1.0",
     "gulp-sourcemaps": "^2.6.5",
